[update] build-utils.mjs  update

+++ build-utils.mjs
@@ -6,15 +6,19 @@ import fs from 'fs';
 
 function buildUtils() {
   try {
-    try {
-      fs.mkdirSync('./.output/server/node_modules/@popperjs');
-    } catch (error) {
-      console.log(error);
+    if (fs.existsSync('./.output/server/node_modules/')) {
+      try {
+        fs.mkdirSync('./.output/server/node_modules/@popperjs');
+      } catch (error) {
+        console.log(error);
+      }
     }
-    fs.renameSync(
-      './.output/server/node_modules/@sxzz/popperjs-es',
-      './.output/server/node_modules/@popperjs/core'
-    );
+    if (fs.existsSync('./.output/server/node_modules/@sxzz/popperjs-es')) {
+      fs.renameSync(
+        './.output/server/node_modules/@sxzz/popperjs-es',
+        './.output/server/node_modules/@popperjs/core'
+      );
+    }
   } catch (error) {
     console.error(error);
   }
